alike.integer {rapportools}	R Documentation

Check integers

Description

This function tests if given variable "appears" to be an integer. To qualify as such, two conditions need to be satisfied: it should be stored as numeric object, and it should pass regular expression test if it consists only of digits.

Usage

alike.integer(x)

Arguments

x
a numeric variable that is to be tested

Value

a logical value that indicates that tested variable "looks like" integer

[Package rapportools version 1.1 Index]