fml {rapportools}	R Documentation

Create Formula from Strings

Description

Takes multiple character arguments as left and right-hand side arguments of a formula, and concatenates them in a single string.

Usage

fml(left, right, join.left = " + ", join.right = " + ")

Arguments

left
a string with left-hand side formula argument

right
a character vector with right-hand side formula arguments

join.left
concatenation string for elements of character vector specified in left

join.right
concatenation string for elements of character vector specified in right

Examples

fml("hp", c("am", "cyl"))    # "hp ~ am + cyl"

[Package rapportools version 1.1 Index]