htest {rapportools}	R Documentation

Hypothesis Tests

Description

This function uses htest.short, to extract statistic and p-value from htest-classed object. Main advantage of using htest is that it's vectorised, and can accept multiple methods.

Usage

htest(
  x,
  ...,
  use.labels = getOption("rapport.use.labels"),
  use.method.names = TRUE,
  colnames = c("Method", "Statistic", "p-value")
)

Arguments

x
arguments to be passed to function specified in test

...
additional arguments for function specified in test

use.labels
a logical value indicating whether variable labels should be placed in row names. If set to FALSE, output of deparse(substitute(x)) will be used.

use.method.names
use the string provided in method attribute of htest object

colnames
a character string containing column names

Details

Default parameters are read from options:

'rapport.use.labels'.

Value

a data.frame with applied tests in rows, and their results (statistic and p-value) in columns

Examples

## Not run: 
library(nortest)
htest(rnorm(100), shapiro.test)
htest(rnorm(100), lillie.test, ad.test, shapiro.test)
htest(mtcars, lillie.test)
htest(mtcars, lillie.test, ad.test, shapiro.test)

## End(Not run)

[Package rapportools version 1.1 Index]