nmissing {rapportools}	R Documentation

Number of Missing Cases

Description

Returns a number of missing (NA) values in a variable. This is a wrapper around univar function with anonymous function passed to count number of NA elements in a variable.

Usage

nmissing(...)

Arguments

...
parameters to be passed to univar function

Value

a numeric value with number of missing vector elements

[Package rapportools version 1.1 Index]