nvalid {rapportools}	R Documentation

Number of Valid Cases

Description

Returns the number of valid (non-NA) values in a variable. This is a wrapper around univar function with length function passed in fn argument, but with missing values previously removed. However, it's not possible to cancel NA omission with this function (doing so will yield error).

Usage

nvalid(...)

Arguments

...
parameters to be passed to univar function

Value

a numeric value with number of valid (non-NA) vector elements

[Package rapportools version 1.1 Index]