strwidthest {rapportools}	R Documentation

Estimate string width

Description

Pre-computed width of strings without actually calling to graphics and new.plot. The function can only handle base ASCII characters and default width of those is estimated by using standard 12 pt serif on a standard plot. Non-ASCII characters are replaced by an underscore.

Usage

strwidthest(s)

Arguments

s
string

Value

numeric value representing the total width of the provided string in millimeters

Examples

## Not run: 
strwidthrel('R') # 71
strwidthrel('R is awesome!') # 635

## End(Not run)

[Package rapportools version 1.1 Index]