synonyms {rapportools}	R Documentation

Get or set synonyms list

Description

Without the l parameter, this function returns the saved list of synonym words. If l is set, then this word list is saved for future use.

Usage

synonyms(l)

Arguments

l
a grouped list of words

Examples

{
synonyms(list(c('package', 'library'), c('foo', 'bar', 'baz')))
synonyms()
}

[Package rapportools version 1.1 Index]