univar {rapportools}	R Documentation

Descriptive Statistics

Description

This function operates only on vectors or their subsets, by calculating a descriptive statistic specified in fn argument.

Usage

univar(x, subset = NULL, fn, na.rm = TRUE, ...)

Arguments

x
a numeric variable to be summarised

subset
an expression that evaluates to logical vector (defaults to NULL, in which case the function specified in fun is applied on a vector)

fn
a function or a function name to be applied on a variable or it's subset

na.rm
a logical value indicating whether NA's should be removed (defaults to TRUE)

...
additional arguments for function specified in fn

Value

a numeric

[Package rapportools version 1.1 Index]